Add switching mpm
This is a little difficult as the ansible apache2 module checks
that one mpm is enabled, and only one. So switching mpm with
the official apache2 ansible module requires

	ignore_configcheck: true
	warn_mpm_absent: false

so we ourselves need to make shure there is an enabled one, and
only one.

So we start by disabling all known mpms we dont want, and then
explicitly enabling the one we want.